packages/pragma/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pragma"
version = "0.4.0"
description = "Senior engineer on rails for AI-driven development"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [{ name = "Pragma contributors" }]
dependencies = [
  "pragma-sdk>=0.4.0",
  "typer>=0.12",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "types-PyYAML>=6.0",
  "opentelemetry-api>=1.30",
  "opentelemetry-sdk>=1.30",
]

[project.scripts]
pragma = "pragma.__main__:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pragma"]

[tool.hatch.build.targets.wheel.force-include]
"src/pragma/templates" = "pragma/templates"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "S", "UP", "SIM", "C90"]
ignore = ["S101", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "C90"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src/pragma"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
